Add note about ingress-nginx
Using ingress-nginx in TLS-terminating mode (the default configuration) appears to work correctly, as I have verified using the test script provided in this repository against a couple of different environments using ingress-nginx. A careful reading of the linked bug seems to indicate that this is only an issue in passthrough mode.